The code used to get results from the YouTube API
NOTE: This file wasn't actually written in Python. It was written using the Apps Script extension for Google Sheets. Apps Script doens't let you download the file that contains the code. We made this file so that we could have a record of the code somewhere besides the Apps Script extension